Refactor adapters
