+++ daiquiri/core/adapter/database/postgres.py
@@ -3,12 +3,12 @@ import warnings
 
 from django.db import OperationalError, ProgrammingError
 
-from .base import DatabaseAdapter
+from .base import BaseDatabaseAdapter
 
 logger = logging.getLogger(__name__)
 
 
-class PostgreSQLAdapter(DatabaseAdapter):
+class PostgreSQLAdapter(BaseDatabaseAdapter):
 
     DATATYPES = {
         'character': {
